bot/requirements.txt
# =====================================================
# Telegram Bot - Python Dependencies
# =====================================================

# Telegram Bot Framework
aiogram==3.13.1

# Database & Auth
supabase==2.9.0
postgrest==0.17.1

# Environment & Config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP Clients
httpx==0.27.2
aiohttp==3.10.5
requests==2.32.3

# Logging
loguru==0.7.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Cache (опционально)
aiocache==0.12.2
redis==5.1.1

# Testing (dev)
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Code Quality (dev)
black==24.8.0
flake8==7.1.1
mypy==1.11.2
isort==5.13.2

